requirements.txt
alembic==1.11.1
anyio==3.6.2
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
Flask==2.3.2
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
python-telegram-bot==20.3
sniffio==1.3.0
SQLAlchemy==2.0.14
typing-extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
pytelegrambotapi==4.11.0
